+++ src/server/Multimedia/MultimediaCustom.ts
@@ -16,21 +16,22 @@ export class MultimediaCustom {
             node.response.end(JSON.stringify(response));
         });
     };
-
+    
     create = (body: any): Promise<any> => {
         const api: ApiModule = new ApiModule(new Multimedia());
-
+        
         return api.create({ body }, {});
     };
-
-    update = (node: iNode) => {
-        let api: ApiModule = new ApiModule(new Multimedia());
-
-        const hooks: any = {
-        };
-
-        api.update({ body: node.request.body, pk: node.request.params }, hooks).then((response) => {
+    
+    updateRequestHandler = (node: iNode) => {
+        this.update(node.request.body, node.request.params).then((response) => {
             node.response.end(JSON.stringify(response));
         });
+    };
+
+    update = (body: any, pk: any): Promise<any> => {
+        let api: ApiModule = new ApiModule(new Multimedia());
+
+        return api.update({ body, pk }, {});
     };
 }